+++ src/FluxPwaGenerator.mjs
@@ -26,12 +26,14 @@ export class FluxPwaGenerator {
     }
 
     /**
+     * @param {string} fallback_icon_template_svg_file
      * @param {string} manifest_json_file
      * @returns {Promise<void>}
      */
-    async generateIcons(manifest_json_file) {
+    async generateIcons(fallback_icon_template_svg_file, manifest_json_file) {
         await (await import("./Pwa/GenerateIcons.mjs")).GenerateIcons.new()
             .generateIcons(
+                fallback_icon_template_svg_file,
                 manifest_json_file
             );
     }